--- src/components/account-settings-overlay/settings-tabs-container/styles.styl
@@ -1,20 +1,0 @@
-.tabsContainer
-  display: flex
-
-.tabs
-  flex-basis: 20%
-  border-right: 1px solid section-line
-  padding: 0 10px 0 0
-  list-style: none
-
-.tab
-  margin-bottom: 5px
-  cursor: pointer
-  display: block
-
-.tabPanels
-  flex-basis: 80%
-  padding-left: 10px
-
-.tabPanelSection
-  padding-bottom: 100px
