+++ src/presenters/pages/settings.styl
@@ -1,20 +1,22 @@
 @import  "../../components/constants"
 
-.settingsTabs
+.tabsContainer
   display: flex
 
-.settingsActions
+.tabs
   flex-basis: 20%
   border-right: 1px solid section-line
   padding: 0 10px 0 0
   list-style: none
 
-.settingsTab
+.tab
   margin-bottom: 5px
   cursor: pointer
+  display: block
 
-.settingsContent
+.tabPanels
   flex-basis: 80%
   padding-left: 10px
-  .contentSection
-    padding-bottom: 100px
+
+.tabPanelSection
+  padding-bottom: 100px
